Do not count delete markers during distinct key calculation in properties collector
Summary:
We should not be including delete markers when calculating number of distinct keys, since they are not visible. Also add an assert to ensure that distinctness is always less than the number of rows in that sst file.

Reference patch: https://github.com/facebook/mysql-5.6/commit/44d6a8df12cf827701a7a5546368db7f88892f04

Reviewed By: yizhang82

Differential Revision: D26283954

+++ storage/rocksdb/properties_collector.h
@@ -74,7 +74,7 @@ struct Rdb_index_stats {
   void merge(const Rdb_index_stats &s, const bool increment = true,
              const int64_t estimated_data_len = 0);
 
-  void adjust_cardinality(uint64_t adjustment_factor);
+  void adjust_cardinality(double adjustment_factor);
 
   void reset_cardinality();
 };
